fix(embedTokens): better api surface, repo structure
