+++ packages/server/modules/core/graph/mocks/core.ts
@@ -10,11 +10,6 @@ const mocks: SpeckleModuleMocksConfig = {
       blob: () => {
         return store.get('BlobMetadata') as any
       }
-    },
-    ProjectMutations: {
-      createEmbedToken: () => {
-        return faker.string.hexadecimal({ length: 42 })
-      }
     }
   }),
   mocks: {
