+++ packages/server/test/graphql/embedTokens.ts
@@ -3,7 +3,9 @@ import gql from 'graphql-tag'
 export const createEmbedTokenMutation = gql`
   mutation CreateEmbedToken($token: EmbedTokenCreateInput!) {
     projectMutations {
-      createEmbedToken(token: $token)
+      createEmbedToken(token: $token) {
+        token
+      }
     }
   }
 `
