--- packages/shared/src/authz/policies/project/tokens/canReadEmbedTokens.ts
@@ -1,54 +1,0 @@
-import { err, ok } from 'true-myth/result'
-import {
-  ProjectNoAccessError,
-  ProjectNotEnoughPermissionsError,
-  ProjectNotFoundError,
-  ServerNoAccessError,
-  ServerNoSessionError,
-  ServerNotEnoughPermissionsError,
-  WorkspaceNoAccessError,
-  WorkspaceNotEnoughPermissionsError,
-  WorkspaceSsoSessionNoAccessError
-} from '../../../domain/authErrors.js'
-import { MaybeUserContext, ProjectContext } from '../../../domain/context.js'
-import { Loaders } from '../../../domain/loaders.js'
-import { AuthPolicy } from '../../../domain/policies.js'
-import { ensureImplicitProjectMemberWithReadAccessFragment } from '../../../fragments/projects.js'
-
-export const canReadEmbedTokensPolicy: AuthPolicy<
-  | typeof Loaders.getProject
-  | typeof Loaders.getEnv
-  | typeof Loaders.getServerRole
-  | typeof Loaders.getWorkspaceRole
-  | typeof Loaders.getWorkspace
-  | typeof Loaders.getWorkspaceSsoProvider
-  | typeof Loaders.getWorkspaceSsoSession
-  | typeof Loaders.getProjectRole
-  | typeof Loaders.getAdminOverrideEnabled,
-  MaybeUserContext & ProjectContext,
-  InstanceType<
-    | typeof ProjectNotFoundError
-    | typeof ServerNoAccessError
-    | typeof ServerNoSessionError
-    | typeof ServerNotEnoughPermissionsError
-    | typeof ProjectNoAccessError
-    | typeof WorkspaceNoAccessError
-    | typeof WorkspaceSsoSessionNoAccessError
-    | typeof WorkspaceNotEnoughPermissionsError
-    | typeof ProjectNotEnoughPermissionsError
-  >
-> =
-  (loaders) =>
-  async ({ userId, projectId }) => {
-    const hasReadAccess = await ensureImplicitProjectMemberWithReadAccessFragment(
-      loaders
-    )({
-      userId,
-      projectId
-    })
-    if (hasReadAccess.isErr) {
-      return err(hasReadAccess.error)
-    }
-
-    return ok()
-  }
